ai-service/requirements.txt
flask
flask-cors
pillow
torch
torchvision
transformers
opencv-python
numpy
google-cloud-vision
